red_agent/requirements.txt
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.11.3
DateTime==5.5
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
elastic-transport==9.2.0
elasticsearch==9.2.0
fastapi==0.121.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
kafka-python==2.0.4
kombu==5.5.4
msgpack==1.1.2
openai==2.7.2
packaging==25.0
pika==1.3.1
prompt_toolkit==3.0.52
py==1.11.0
pydantic==2.12.4
pydantic_core==2.41.5
pymetasploit3==1.0.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
retry==0.9.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1
